Update group_icon_repos.json
Its Mila Martin.  This is a new token i made that will only be for quality apng animation in the juungle.  The files are highly compressed with small pixel dimensions.  they are generally a 2 second loop that is 200x200 allowing the file size to stay very small.  the 32/64/128 files are still photos and the original file is an apng file that i will make sure to keep small.   purchasers of these tokens will also be given a private download address to my own site/hosting where they can download a much larger higher quality gif of the animation if they want. that way its not on other peoples servers, just my own.   I want to bring quality animation to the juungle and create a new visually exciting line of NFTs for juungle users to enjoy and to help make juungle a place that has cool animated NFTs on it.