Update verified_tokens.json
hello.  this is Mila Martin.  Ive created a regular slp token (non nft) that will accompany my new animated nft token.  this regular token will be used as a rewards device. it will be sent out to holders of my nfts on a monthly basis and they can redeem this token for exclusive special nfts available only on my website, and only by purchasing them with this token.  i submitted the token icon to a completely different repository but i would also like the verified check mark on it too.  im not sure if this repo and file are only for nfts or also regular tokens but i figured id try to get the check mark here.  thank you for your time.